requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
greenlet==3.1.1
idna==3.10
kagglehub==0.3.6
numpy==2.2.2
packaging==24.2
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
pytz==2024.2
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.37
sqlalchemy-cockroachdb==2.0.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
Unidecode==1.3.8
urllib3==2.3.0
